+++ elide-datastore/elide-datastore-aggregation/src/test/resources/configs/models/security.hjson
