--- elide-datastore/elide-datastore-aggregation/src/test/resources/configs_with_security_hjson/db/sql/SalesDBConnection.hjson
@@ -1,16 +1,0 @@
-{
-  dbconfigs:
-  [
-    {
-      name: SalesDBConnection
-      url: jdbc:h2:mem:SalesDB;DB_CLOSE_DELAY=-1;INIT=RUNSCRIPT FROM 'classpath:prepare_SalesDB_tables.sql'
-      driver: org.h2.Driver
-      user: guest
-      dialect: H2
-      propertyMap:
-      {
-        encrypted.password: Z3Vlc3RQYXNzd29yZA==
-      }
-    }
-  ]
-}
